refactor client resource files
